feat: recurse sketches' path
includes a simple self test of the action that is triggered manually or on a PR.
