+++ tests/test_assets/out.md
@@ -4,6 +4,7 @@
 | Board | Flash | % | RAM for global variables | % |
 |---|---|---|---|---|
 | arduino:avr:leonardo | :green_heart: -40 - 0 | -0.14 - 0 | :green_heart: -16 - 0 | -0.62 - 0 |
+| arduino:avr:nano |  0 - 0 | 0 - 0 |  0 - 0 | 0 - 0 |
 | arduino:sam:arduino_due_x |  0 - 0 | 0 - 0 | N/A | N/A |
 | arduino:samd:mkrzero | :grey_question: -28 - +84 | -0.01 - +0.03 | :small_red_triangle: 0 - +32 | 0 - +0.1 |
 
@@ -17,6 +18,17 @@
 | examples/MIDIUSB_loop | 0 | 0 | 0 | 0 |
 | examples/MIDIUSB_read | 0 | 0 | 0 | 0 |
 | examples/MIDIUSB_write | -40 | -0.14 | -16 | -0.62 |
+
+### `arduino:avr:nano`
+
+| Sketch | Flash | % | RAM for global variables | % |
+|---|---|---|---|---|
+| examples/helloworld_rx | 0 | 0 | 0 | 0 |
+| examples/helloworld_rx_advanced | 0 | 0 | 0 | 0 |
+| examples/helloworld_tx_advanced | 0 | 0 | 0 | 0 |
+| examples/helloworld_tx | 0 | 0 | 0 | 0 |
+| examples/Network_Priority_RX | 0 | 0 | 0 | 0 |
+| examples/Network_Priority_TX | 0 | 0 | 0 | 0 |
 
 ### `arduino:sam:arduino_due_x`
